ASP.NET Core RC2 Massive Upgrade
First and foremost, the major segments of this commit is to address the
package differences with RC1 here and there.

Apart from that, here are the list of changes:

BrandCategory Web API Controller
- Introduce a new Get + BrandId method to pull specific Ids a Brand has.

Brands/Update.cshtml
- Fixes made to the existing magicsuggest box.
Previously when the box is loaded, it is unable to load the selected
categories properly from time to time. The new fix is to prevent the
variable where I store my pulled data to sync with the magicsuggest box
immediately. Instead, I force the variable to fully GET the data first
before I set the selections to the magic suggest box.