Products: PUT: Transistor Method
Unlike the previous logic where we use two columns, MetricAmount and MetricType as Composite Keys to identify whether if it's an existing metric and then proceed on to create, update or delete it.

PUT will now utilize the Transistor Method I've thought of. Each existing Metric will be preloaded into the Update View with it's corresponding MetricId. Should the Metric be a new Metric, It's MetricId value will be 0. Thereafter, We'll change and auto generate it's own MetricId.